requirements.txt
wandb==0.13.9
torchvision==0.15.2
torch==2.0.0
Pillow==9.4.0
imageio==2.22.3